# ex-04-feign/services/store-api/src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Operations for Books.",
    "version" : "99",
    "title" : "Book API"
  },
  "host" : "localhost:8081",
  "basePath" : "/book-api/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/auth" : {
      "post" : {
        "operationId" : "token",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/authors" : {
      "get" : {
        "tags" : [ "authors" ],
        "summary" : "Find all authors",
        "description" : "",
        "operationId" : "getAllAuthors",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "All authors found",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/AuthorDTO"
              }
            }
          },
          "404" : {
            "description" : "Authors not found"
          }
        }
      },
      "post" : {
        "tags" : [ "authors" ],
        "summary" : "Create an author",
        "description" : "",
        "operationId" : "createAuthor",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Author to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AuthorDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AuthorDTO"
            }
          },
          "201" : {
            "description" : "The author is created"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "415" : {
            "description" : "Format is not JSon"
          }
        }
      },
      "put" : {
        "tags" : [ "authors" ],
        "summary" : "Update an author",
        "description" : "",
        "operationId" : "updateAuthor",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Author to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AuthorDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The author is updated",
            "schema" : {
              "$ref" : "#/definitions/AuthorDTO"
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/authors/{id}" : {
      "get" : {
        "tags" : [ "authors" ],
        "summary" : "Find an author by the Id.",
        "description" : "",
        "operationId" : "getAuthor",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Author found",
            "schema" : {
              "$ref" : "#/definitions/AuthorDTO"
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Author not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "authors" ],
        "summary" : "Delete an author",
        "description" : "",
        "operationId" : "deleteAuthor",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Author has been deleted"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
...
    "/categories" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Find all authors",
        "description" : "",
        "operationId" : "getAllCategories",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "All categories found",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CategoryDTO"
              }
            }
          },
          "404" : {
            "description" : "Categories not found"
          }
        }
      },
      "post" : {
        "tags" : [ "categories" ],
        "summary" : "Create a category",
        "description" : "",
        "operationId" : "createCategory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Category to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CategoryDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CategoryDTO"
            }
          },
          "201" : {
            "description" : "The category is created"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "415" : {
            "description" : "Format is not JSon"
          }
        }
      },
      "put" : {
        "tags" : [ "categories" ],
        "summary" : "Update a category",
        "description" : "",
        "operationId" : "updateCategory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Category to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CategoryDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The category is updated",
            "schema" : {
              "$ref" : "#/definitions/CategoryDTO"
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/categories/{id}" : {
      "get" : {
        "tags" : [ "categories" ],
        "summary" : "Find a category by the Id.",
        "description" : "",
        "operationId" : "getCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Category found",
            "schema" : {
              "$ref" : "#/definitions/CategoryDTO"
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "404" : {
            "description" : "Category not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "categories" ],
        "summary" : "Delete a category",
        "description" : "",
        "operationId" : "deleteCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Category has been deleted"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "AuthorDTO" : {
      "type" : "object",
      "required" : [ "firstName", "lastName" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "firstName" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 50
        },
        "lastName" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 50
        },
        "bio" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 5000
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date"
        },
        "preferredLanguage" : {
          "type" : "string",
          "enum" : [ "ENGLISH", "FRENCH", "SPANISH", "PORTUGUESE", "ITALIAN", "FINISH", "GERMAN", "DEUTSCH", "RUSSIAN" ]
        }
      }
    },
...
    "CategoryDTO" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string",
          "minLength" : 2,
          "maxLength" : 100
        }
      }
    },
...
  }
}